requirements.txt
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Faker==37.4.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
idna==3.10
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
natsort==8.4.0
outcome==1.3.0.post0
overrides==7.7.0
ply==3.11
protobuf==5.29.3
pycparser==2.22
pymongo==4.13.1
PySocks==1.7.1
referencing==0.36.2
requests==2.32.3
robotframework==7.2.2
robotframework-assertion-engine==3.0.3
robotframework-browser==19.5.1
robotframework-faker==5.0.0
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.7.1
rpds-py==0.25.1
seedir==0.5.1
selenium==4.34.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
